requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==5.0.2
django-crispy-forms==2.1
django-widget-tweaks==1.5.0
mysqlclient==2.2.3
packaging==23.2
plotly==5.18.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.9.0
